refactoring on criteria
