+++ public/sw.js
@@ -32,12 +32,7 @@ self.addEventListener(`install`, (evt) => {
           `/img/icons/sightseeing.png`,
           `/img/icons/taxi.png`,
           `/img/icons/train.png`,
-          `/img/icons/transport.png`,
-          // `/img/photos/1.jpg`,
-          // `/img/photos/2.jpg`,
-          // `/img/photos/3.jpg`,
-          // `/img/photos/4.jpg`,
-          // `/img/photos/5.jpg`,
+          `/img/icons/transport.png`
         ]));
       })
   );
