+++ src/view/abstract.js
@@ -26,6 +26,7 @@ export default class Abstract {
   removeElement() {
     this._element = null;
   }
+
   show() {
     this.getElement().classList.remove(`visually-hidden`);
   }
