+++ src/view/smart.js
@@ -25,8 +25,9 @@ export default class Smart extends Abstract {
   }
 
   updateElement() {
-    let prevElement = this.getElement();
+    const prevElement = this.getElement();
     const parent = prevElement.parentElement;
+
     this.removeElement();
 
     const newElement = this.getElement();
